Replace pytz with zoneinfo
This commit replaces the external pytz library with zoneinfo, which
provides the same functionality as pytz but as part of the standard
library.
This change is partially inspired by [Debian's](https://packages.debian.org/sid/python3-pytz-deprecation-shim) comments on how
the library has a non-standard interface which tends to be error-prone.
However, it is still one less dependency to worry about.