Fix: Broken locales after installation
This commit solves the issue where locales would be generated but not
properly recognized by the system after installation. We solve this by
bind mounting a new locales directory, stored in the mutable part of the
system (/var/lib/abroot/etc/vos-a/locales), into the original locales
directory in /usr/lib. This mount is done in the same way as /home and
/opt, which are mounted via Systemd unit.